Refactors routes into their specific files
