+++ src/routes/index.js
@@ -1,0 +1,13 @@
+import SearchStack from './search';
+import NotifStack from './notification';
+import ProfileStack from './profile';
+import FilterStack from './filter';
+import Tabs from './tabs';
+
+export default {
+  SearchStack,
+  NotifStack,
+  ProfileStack,
+  FilterStack,
+  Tabs,
+};
