+++ src/routes/tabs.js
@@ -1,0 +1,43 @@
+import { TabNavigator } from 'react-navigation';
+import HomeScreen from '../screens/HomeScreen';
+import SearchStack from './search';
+import NotifStack from './notification';
+import ProfileStack from './profile';
+
+import { tabRed, listBackPurple } from '../constants/colors';
+
+const Tabs = TabNavigator(
+  {
+    Home: {
+      screen: HomeScreen,
+    },
+    Search: {
+      screen: SearchStack,
+    },
+    Profile: {
+      screen: ProfileStack,
+    },
+    Feed: {
+      screen: HomeScreen,
+    },
+    Notif: {
+      screen: NotifStack,
+    },
+  },
+  {
+    tabBarOptions: {
+      activeTintColor: tabRed,
+      inactiveBackgroundColor: listBackPurple,
+      activeBackgroundColor: listBackPurple,
+      showLabel: false,
+      lazy: true,
+      style: {
+        height: 44.96,
+        borderTopWidth: 0,
+      },
+      backgroundColor: listBackPurple,
+    },
+  },
+);
+
+export default Tabs;
